add migration modal fix
